move theLoop of of robot folder
