+++ server/modules/theLoop.js
@@ -1,8 +1,8 @@
-const pool = require('../pool');
-const authedClient = require('../authedClient');
-const databaseClient = require('../databaseClient/databaseClient');
-const socketClient = require('../socketClient');
-const robot = require('./robot');
+const pool = require('./pool');
+const authedClient = require('./authedClient');
+const databaseClient = require('./databaseClient/databaseClient');
+const socketClient = require('./socketClient');
+const robot = require('./robot/robot');
 
 // set up side toggle - can be boolean, may as well be buys first since trade-pairs are always buys first
 // needs to be outside the loop so it can persist
